- Refactor out/Rename the StepOne and StepTwo components
